requirements.txt
aiohttp==3.9.5  # Required for azure-search-documents, azure-ai-contentsafety, azure-cosmos
aiosqlite==0.20.0
azure-ai-contentsafety==1.0.0
azure-ai-translation-text==1.0.0b1
azure-communication-callautomation==1.2.0
azure-communication-sms==1.0.1
azure-cosmos==4.6.0
azure-eventgrid==4.19.0
azure-identity==1.16.0  # Required for all Azure SDKs
azure-monitor-opentelemetry==1.4.1
azure-search-documents==11.6.0b3
fastapi==0.111.0
gunicorn==22.0.0
jinja2==3.1.3
json-repair==0.17.0
mistune==3.0.2
openai==1.25.2
opentelemetry-instrumentation-aiohttp-client==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-openai==0.17.1
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-sqlite3==0.45b0
phonenumbers==8.13.36  # Required for pydantic-extra-types
pydantic-extra-types==2.7.0  # Required for pydantic
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
redis==5.0.4
tenacity==8.2.3
tiktoken==0.6.0
twilio==9.0.5
typing-extensions==4.11.0
uvicorn[standard]==0.29.0  # Required for Gunicorn
